+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added more examples and documentation
 - Added `DezignableBackground` with background fill color and gradient background
 - Added `DezignableShadow`
+- Added component `DezignableLabel`
 
 ## 1.1.0 - 26 feb 2016
 - Mostly refactored to protocol based extensions